requirements.txt
streamlit
google-generativeai
fpdf
chromadb
pysqlite3-binary